requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
python-multipart==0.0.9
httpx==0.27.2
PyPDF2==3.0.1
chromadb==1.5.5
